little changes

+++ lib/src/com/n8cats/lib_gwt/LibAllGwt.java
@@ -217,13 +217,13 @@ public static class Fun {
 	}
 
 	public static int sign(float a) {
-		return sign((double)a);
+		return sign((double) a);
 	}
 	public static int sign(double a) {
 		if(a == 0) {
 			return 0;
 		}
-		return (int) (a/ java.lang.Math.abs(a));
+		return (int) (a / java.lang.Math.abs(a));
 	}
 	public static double positive(double v) {
 		if(v > 0) {
@@ -232,10 +232,10 @@ public static class Fun {
 		return 0;
 	}
 	public static float positive(float v) {
-		return (float)positive((double)v);
+		return (float) positive((double) v);
 	}
-	public static float arg0toInf(float y, float middle) {
-		return y/middle/(1+y/middle);
+	public static float arg0toInf(double y, float middle) {
+		return (float) (y / middle / (1 + y / middle));
 	}
 }
 }
